Add `Timer.recordNanoseconds` Generic Overload (#28)
Motivation:

As discussed in Issue #28, there are a few preferred lightweight APIs for getting timestamps such as `DispatchTime` but those APIs return `UInt64` types, which need to be handled for overflow.

Modifications:

Added a generic `BinaryInteger` overload method for `recordNanoseconds` that guards against accidental overflows from `UInt64` to `Int64`

Result:

Users now have access to a new method to record nanosecond measurements with `Timer` that guards against overflows.

+++ Sources/Metrics/Metrics.swift
@@ -52,7 +52,7 @@ public extension Timer {
     func record(_ duration: DispatchTimeInterval) {
         switch duration {
         case .nanoseconds(let value):
-            self.recordNanoseconds(Int64(value))
+            self.recordNanoseconds(value)
         case .microseconds(let value):
             self.recordMicroseconds(value)
         case .milliseconds(let value):
